pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-to-confluence"
version = "0.1.0"
description = "Convert markdown files with mermaid diagrams to Confluence pages"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.0",
]

[project.scripts]
markdown-to-confluence = "markdown_to_confluence.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["markdown_to_confluence*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
